feat: interating on SiteConfig and ModuleConfig interfaces
The first is to build a site, the second is to independently build a set of modules for module federation.
