+++ shell/test.site.config.tsx
@@ -1,6 +1,7 @@
 import { SiteConfig } from '..';
 
 const config: SiteConfig = {
+  apps: [],
   ACCOUNT_PROFILE_URL: 'http://localhost:1995',
   ACCOUNT_SETTINGS_URL: 'http://localhost:1997',
   BASE_URL: 'http://localhost:8080',
